# specs/002-diagnostico-contexto/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: "White Tree Nexus — Diagnóstico e Contexto (ISO 27001 Cláusula 4)"
  version: 0.1.0
  description: >
    Contratos da feature 002 (Cláusula 4): diagnóstico, análise de contexto (4.1), mapa de partes
    interessadas (4.2), declaração de escopo (4.3), visão consolidada/sugestões e política de
    classificação. Sobre a fundação multi-tenant. NÃO normativo de stack.
servers:
  - url: /
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  parameters:
    OrgContext:
      name: X-Org-Context
      in: header
      required: false
      schema: { type: string, format: uuid }
      description: "Organização-alvo (obrigatório p/ Consultor multi-org e Super Admin)."
  responses:
...
    Forbidden: { description: "Sem permissão no próprio tenant.", content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    NotFound: { description: "Recurso inexistente OU de outro tenant (genérico).", content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...
  schemas:
    Error:
      type: object
      properties: { detail: { type: string } }
      required: [detail]
    DocStatus: { type: string, enum: [draft, in_review, in_force, obsolete] }
    Classification: { type: string, enum: [publico, uso_interno, confidencial, restrito] }
    Level: { type: string, enum: [alto, medio, baixo] }
    EngagementStrategy: { type: string, enum: [manage_closely, keep_satisfied, keep_informed, monitor] }
...
    Stakeholder:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        type: { type: string, enum: [internal, external] }
        power: { $ref: '#/components/schemas/Level' }
        interest: { $ref: '#/components/schemas/Level' }
        strategy: { $ref: '#/components/schemas/EngagementStrategy' }
    ScopeItem:
      type: object
      properties:
        id: { type: string, format: uuid }
        kind: { type: string, enum: [inclusion, exclusion] }
        description: { type: string }
        justification: { type: string }
    DocumentVersion:
      type: object
      properties:
        id: { type: string, format: uuid }
        identifier: { type: string }
        version_number: { type: integer }
        status: { $ref: '#/components/schemas/DocStatus' }
        classification: { $ref: '#/components/schemas/Classification' }
        emitted_at: { type: string, format: date-time }
        next_review_at: { type: string, format: date-time }
        approved_by: { type: string, format: uuid }
security:
  - bearerAuth: []
paths:
...
  /context/stakeholders:
    get:
      summary: "Mapa de Partes Interessadas (4.2) — view_context"
      parameters: [{ $ref: '#/components/parameters/OrgContext' }]
      responses: { '200': { description: "OK", content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Stakeholder' } } } } } }
    post:
      summary: "Adicionar parte interessada (estratégia derivada de Poder×Interesse) — manage_context"
      parameters: [{ $ref: '#/components/parameters/OrgContext' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/Stakeholder' } } } }
      responses: { '201': { description: "Criada (com strategy derivada)", content: { application/json: { schema: { $ref: '#/components/schemas/Stakeholder' } } } }, '403': { $ref: '#/components/responses/Forbidden' } }
  /context/stakeholders/{stakeholder_id}:
    patch:
      summary: "Editar parte interessada (re-deriva strategy) — manage_context"
      parameters: [{ name: stakeholder_id, in: path, required: true, schema: { type: string, format: uuid } }, { $ref: '#/components/parameters/OrgContext' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/Stakeholder' } } } }
      responses: { '200': { description: "OK", content: { application/json: { schema: { $ref: '#/components/schemas/Stakeholder' } } } }, '404': { $ref: '#/components/responses/NotFound' } }
    delete:
      summary: "Remover parte interessada — manage_context"
      parameters: [{ name: stakeholder_id, in: path, required: true, schema: { type: string, format: uuid } }, { $ref: '#/components/parameters/OrgContext' }]
      responses: { '204': { description: "Removida" }, '404': { $ref: '#/components/responses/NotFound' } }
  /context/stakeholders/submit-review:
    post:
      summary: "Enviar Mapa para revisão — manage_context"
      parameters: [{ $ref: '#/components/parameters/OrgContext' }]
      responses: { '200': { description: "Em revisão", content: { application/json: { schema: { type: object } } } } }
  /context/stakeholders/approve:
    post:
      summary: "Aprovar/emitir versão do Mapa — approve_context_document"
      parameters: [{ $ref: '#/components/parameters/OrgContext' }]
      requestBody: { required: true, content: { application/json: { schema: { type: object } } } }
      responses: { '201': { description: "Versão emitida", content: { application/json: { schema: { $ref: '#/components/schemas/DocumentVersion' } } } }, '403': { $ref: '#/components/responses/Forbidden' } }
  /context/scope:
    get:
      summary: "Declaração de Escopo (4.3) — view_context"
      parameters: [{ $ref: '#/components/parameters/OrgContext' }]
      responses: { '200': { description: "OK", content: { application/json: { schema: { type: object } } } } }
    put:
      summary: "Editar interfaces/dependências e referências de versão (4.3) — manage_context"
      parameters: [{ $ref: '#/components/parameters/OrgContext' }]
      requestBody: { required: true, content: { application/json: { schema: { type: object, properties: { interfaces_dependencies: { type: string }, context_version_ref: { type: string, format: uuid }, stakeholder_version_ref: { type: string, format: uuid } } } } } }
      responses: { '200': { description: "OK", content: { application/json: { schema: { type: object } } } } }
  /context/scope/items:
    post:
      summary: "Adicionar inclusão/exclusão de escopo (justificada) — manage_context"
      parameters: [{ $ref: '#/components/parameters/OrgContext' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/ScopeItem' } } } }
      responses: { '201': { description: "Criado", content: { application/json: { schema: { $ref: '#/components/schemas/ScopeItem' } } } } }
  /context/scope/approve:
    post:
      summary: "Aprovar/emitir versão da Declaração de Escopo — approve_context_document"
      parameters: [{ $ref: '#/components/parameters/OrgContext' }]
      requestBody: { required: true, content: { application/json: { schema: { type: object } } } }
      responses:
        '201': { description: "Versão emitida", content: { application/json: { schema: { $ref: '#/components/schemas/DocumentVersion' } } } }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { description: "Conflito (ex.: referências de versão ausentes/obsoletas).", content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...